:recycle: [REFACTOR] 친구피드 탭 클릭 시 api 수정
컨트롤러의 코드를 service로 이동시킴
