+++ BackEnd/favshare/src/main/java/com/favshare/pop/controller/PopController.java
@@ -137,63 +137,10 @@ public class PopController {
 	@GetMapping("/friend")
 	public ResponseEntity<List<FriendFeedDto>> showFriendList(@RequestBody UserIdRequest userIdRequest) {
 		int userId = userIdRequest.getUserId();
-
-		// 친구숫자 세기 (친구: 내가 팔로우 한 사람)
-		List<FriendFeedDto> friendsPopList = userService.getFollowingList(userId);
-
-		if (friendsPopList.size() == 0) { // 친구가 없으면 -> 알고리즘으로 뿌려주기
-			List<FriendFeedDto> result = new ArrayList<FriendFeedDto>();
-
-			// 1. 나를 팔로우 한 사람
-			List<UserProfileDto> UserProfileDtoList = userService.getFollowerList(userId);
-			for (int i = 0; i < UserProfileDtoList.size(); i++) {
-				int id = UserProfileDtoList.get(i).getId();
-				// id로 이사람의 userProfileDto 가져오기
-				UserProfileDto userProfileDto = userService.getUserProfileById(id);
-				// id로 이사람의 popDtoList가져오기
-				List<PopDto> popDtoList = popService.popDtoListByUserId(id);
-				// popDtoList를 for문으로 돌면서
-				for (int j = 0; j < popDtoList.size(); j++) {
-					// new FriendFeedDto한걸 result에 넣어주기
-					result.add(new FriendFeedDto(userProfileDto, popDtoList.get(j)));
-				}
-			}
-
-			// 3. 관심사가 비슷한 사람
-			// 내 관심사
-			List<IdolDto> myInterestIdolList = userService.getInterestIdolList(userId);
-
-			// 나와 관심사가 동일한 사람
-			HashSet<Integer> similarPersonId = new HashSet<Integer>();
-
-
-
-			Iterator<Integer> iter = similarPersonId.iterator();
-			while (iter.hasNext()) {
-				int id = iter.next();
-				// id로 이사람의 userProfileDto 가져오기
-				UserProfileDto userProfileDto = userService.getUserProfileById(id);
-				// id로 이사람의 popDtoList가져오기
-				List<PopDto> popDtoList = popService.popDtoListByUserId(id);
-				// popDtoList를 for문으로 돌면서
-				for (int j = 0; j < popDtoList.size(); j++) {
-					// new FriendFeedDto한걸 result에 넣어주기
-					result.add(new FriendFeedDto(userProfileDto, popDtoList.get(j)));
-				}
-
-			}
-			return new ResponseEntity<List<FriendFeedDto>>(result, HttpStatus.OK);
-
-			// 4. 랜덤
-		} else { // 친구가 있으면 -> 내 친구들의 팝들 최신순으로 보여주기
-
-			friendsPopList.sort(new Comparator<FriendFeedDto>() {
-				@Override
-				public int compare(FriendFeedDto o1, FriendFeedDto o2) {
-					return o2.getPopDto().getCreateDate().compareTo(o1.getPopDto().getCreateDate());
-				}
-			});
-			return new ResponseEntity<List<FriendFeedDto>>(friendsPopList, HttpStatus.OK);
+		try {
+			return ResponseEntity.status(HttpStatus.OK).body(popService.showFriendList(userId));
+		} catch (Exception e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 		}
 
 	}
